+++ src/com/xuxiaocheng/wlist/api/core/storages/exceptions/InvalidStorageConfigException.java
@@ -1,6 +1,8 @@
 package com.xuxiaocheng.wlist.api.core.storages.exceptions;
 
 import java.io.Serial;
+import java.util.Collections;
+import java.util.Map;
 
 /**
  * Thrown when the storage config is invalid.
@@ -10,24 +12,32 @@ public class InvalidStorageConfigException extends RuntimeException {
     private static final long serialVersionUID = -320110429948047601L;
 
     /**
-     * The field that is invalid.
+     * The invalid field map.
+     * The key is the name of the invalid field.
+     * The value is the error message.
      */
-    protected final String field;
+    protected final Map<String, String> messages;
+
+    /**
+     * The extra field name, the key of a global error message.
+     */
+    public static final String Extra = "extra";
 
     /**
      * Internal constructor.
-     * @param field the field that is invalid.
+     * @param messages the field that is invalid.
      */
-    protected InvalidStorageConfigException(final String field) {
-        super("Invalid: " + field);
-        this.field = field;
+    protected InvalidStorageConfigException(final Map<String, String> messages) {
+        super("Invalid: " + messages);
+        this.messages = Collections.unmodifiableMap(messages);
     }
 
     /**
-     * Get the field name that is invalid.
-     * @return the field name.
+     * Get the invalid field map.
+     * @return the invalid field map.
+     * @see #messages
      */
-    public String getField() {
-        return this.field;
+    public Map<String, String> getMessages() {
+        return this.messages;
     }
 }